fix for payments made via /admin
orders placed via /admin were being charged twice.

See these two lines from Spree::Admin::PaymentsController:

    while @order.next; end
    @payment.process! if @order.completed?

For incomplete orders the first line will trigger callbacks that cause
the payment to be charged if the order reaches the "completed" state.
Then the second line will process the payment *again*, resulting in a
double charge.

We can't just remove the second line because orders that are already
complete won't trigger any callbacks from the first line, so these
orders need to have their payments charged manually.

The effect of the bug in our case was that Braintree rejected the second
(duplicate) billing on the payment, and this resulted in the Payment
object being marked as failed when actually it succeeded on the first
attempt. So it appeared from within Spree that a successful payment
was a failed payment.

Fixes #4768


Former-commit-id: efd3c150b1b9b373e992123c6330153e2245b604 [formerly fd10da9a81edca73bcacd0876536e2ff5e9f6ad7 [formerly 79f00b19fa00f9937478a475d6dcefc8c35f84ae]]
Former-commit-id: 511a707423ef966a10cb186e1cfad471ad383864
Former-commit-id: a3378d0dd6638c54a71f8ddcce961ae552ffe4cb [formerly 1341d3aab5646493ccf09d2f0b83573ca02524c5]
Former-commit-id: 66cf4a3d9ae800b202de52eea2e25a468a1cf39b