AvailabilityValidator operates in two modes 1) validate on a single line item 2) validate on a complete order
Both cases check for stock availability in the complete order.

Add stock item class OrderItemBuilder. Its job is to flatten an order with parts to
a list of items. Used by AvailabilityValidator.

Tests: It is important to reload the association `order.line_items` if line_item
is created explicitly as the validator loads it once already.

Former-commit-id: 78221fba6625dec9f97d9065ea12a43fa521294e [formerly c4eceea4719c8ffc5d21bc3461e9ff2486e88b72 [formerly 61b09095bc282d36dcd51258e2692c831d8fe8f8]]
Former-commit-id: fc7db907ecfc2158e47df2b22d15e922de11f43b
Former-commit-id: 0d019d3283b7b413e02c3c20520ef3c98d634b0b [formerly fa2162b5551c67dd1fa43cf4aa7b111bf1f22f5f]
Former-commit-id: 3887ddcb6cc05e441a288ac358207d8196330030