Toggle images when selecting a colour but not a size
This is a (nasty) bugfix to update the images when selecting a different
colour for a RTW item but not selecting the size. Previously it would
not update the image because no variant had been selected. This fix
finds any variant with that colour and uses that to update the images.


Former-commit-id: 1183c5f5223686ffd717b2d91446bd851e5e1a52 [formerly 7c62ed5d1bd273e101e16486803db42143d7ca13]
Former-commit-id: f4d77f851f8e48fd280125e50f2beeb75b58c74c