Optimize Promotion#eligible_rules.
The problem here is that this method was running the same query to load
promotion rules multiple times. Whilst each individual query is fast,
running this query over and over again caused many round trips to the
database in our performance profiling. By mapping eligibility once and
storing it in a local variable we save a lot of that effort.

Fixes #6200


Former-commit-id: 94043df1df02a8c1598c1d772e21898797219a65 [formerly 8dfefc5b497a993e0b0dfa13a50cb20eadb1d9ec]
Former-commit-id: c2841de164d96bd99ed94d55e2a3ea3edab25199